requirements.txt

# ---- Core Web Framework ----
Flask==3.0.3      #stable version 

# ---- Machine Learning / Data Processing csv----
pandas==2.2.2              #stable version
numpy==1.26.4                  
scikit-learn==1.5.0

# ---- Data Loading ----
h5py

# ---- Visualization (useful for analytics reports) ----
matplotlib
seaborn

# --- For image processing --- 
pillow             

# ---- Environment Variables( Always run in venv ----
python-dotenv

